PC-325 Integration Tests - More CreatePatientTests, general patient creation tests
PC-375: Added a helper method to delete all patients from the table on admin view.

PC-375: Added selectors and methods for modifying visibility and patient permissions.

PC-375: Added a publci visibility test. Currently conflicts with the verifyEmailNotifications() test. Probably just need to move to a new class. Will ask about matches not showing up fast enough.

PC-375: Added annotations to ignore deletion of created users. Mdoularized each test to run independently or if needed, with the dependsOnMethods annotation.

PC-375: Added test for collaboration visibility. Asserts that the added user will be able to see the specified user.

PC-375: Changed currentPage variables to something more descriptive. Code refactoring.

PC-325: Added methods to navigate to the Admin's "Refresh matches" page. This is used in the test refreshMatchesForTwoPatients() which goes into the refresh matches page, updates the matches, and asserts that two new ones are found. The assertion will fail currently because matches require a server restart to appear.

PC-325: Added method to remove a collaborator from a patient. removeNthCollaborator() is used in the test to reset to a previous state.

PC-325: Assigned priorities to each test so that they run in order as needed.