PC-325 Integration Tests - Tree traversals on Patient Form
PC-325: Added a test to cycle through the options for creating a patient. Will continue tomorrow.

PC-325: Pre-order traversal of the age of onset buttons added. Added method to allow for the Indication For Referral box to be set. Improved the cycling test and added some more comments.

PC-375: Fixed the publicVisibilityTest and made the page for Refresh Matches actually refresh before checking values within the table.

PC-325: Helper function for a pre-order traversal. Realized that the traverse is only finding the radio button so it cannot populate an array of Strings to represent the labels of each button. Have to do some more digging about how the tree is traversed. Hence there is commented out code in preOrderTraverseAndClick(). The end goal is to have it return the labels of the nodes visited so that the List can be asserted for correctness during a test. Will continue tomorrow.

PC-325: Actually, there maybe nothing wrong with the traverse function. I just passed the wrong child label path to it. Seems to work now. Uncommented the code in the preOrderTraverseAndClick and just changed the child selector path where I concatenated a "> label".

PC-325: Added methods to modify the date of death on the create patient page. Made the cycling test assert that the array of labels is equal to what is expected to be there during a pre-order traversal.