PC-325 Integration Tests - Pedigree Editor Tests and interaction
PC-325: Added methods to navigate to the Pedigree editor from the create patient page. Added method to handle the save warning dialogue box when trying to close the editor.

Pc-325: Added methods for retrieving basic data that is present within the pedigree editor. Ex. Gender of patient. Will need to add more methods to interact with individuals within the editor.

PC-325: Added more methods to get the entered patient info from the pedigree editor.

PC-325: Got the click on the patient edit modal to finally work using an Action class. Need to do some more reading on it as SVGs apparently aren't directly interactable as I thought. Also created a new class for the pedigree page tests, it currently contains a basic one for me to check if navigation to editor works so far.

PC-325: Added method to link to a patient via the pedigree editor. Still working on tree traversal logic. Not working yet. Push this for now.

PC-325: Added methods to retrieve the number of various nodes within the tree. Created a simple test for creating a child when there is one patient present. Need to make methods more robust, abstract away hovering to helper functions, and clean up some commented code.

PC-325: Added methods to create sibling and one for partnership. Selectors are still very unstable. Need to work out why that is, js is causing signficant changes to make it difficult to locate the one selector that is appropriate. Trying to think about logic to make it work. Will have to continue on Monday.