PC-325 Integration Tests - Diagnosis section patient form tests - Slight modification of input traversals in other sections
PC-325 Inserted additional label in inheritance traversal.
- This is a missing label that needed to be asserted for during the test. See PT-3935

PC-325 Added remaining boxes within Prenatal History section in the traversal test.
These boxes are clicked and text inputted.

PC-325 Added methods to interact with the Diagnosis section on the Patient Form.
- Methods and selectors for each input box within the Diagnosis section
- Small test case for checking all input boxes. Will continue on with basic error checking.

PC-325 Inserted more assertions for the visibility of the PubMed boxes when the Case Solved checkbox is toggled.

PC-325 Modified the method that checks for visiblity of the PubMed boxes.
- Changed so that it checks for clickability as the hidden property is still considered visible to the automation framework.
- Added a test case to check for the error messages upon invalid PubMed ID input.

PC-325 Added test for diagnosis section on patient form.
- Methods to extract information from diagnosis section from View Patient page.
- Check error messages upon entering invalid pubMed ID