PC-325 Integration Tests - More matching test cases
PC-325 Added test cases for matching with only genotype and one with only phenotype.
- Overrode navigateToHomePage on the EmailUIPage to directly navigate to the homepage  URL rather than click on PC logo on top left

PC-325 Changed the gene in matchPhenotypes only so that all genotypes are different in that case.
- Added priorities to each test so that they run in a deterministic order.

PC-325: Added methods to move average score and genotype score sliders to their minimum values (Slide all the way to the left)
- Incorporated the genotype slider to slide all the way to the left for phenotype only match test

PC-325 Method that returns a bool to whether or not a match for a given two patients appear on the match table.
- Test case for a user to be presented with error page when trying to access another user's matchable patient