PC-325 Integration Tests - Webdriver is now static
PC-325 Added an xml for All Test cases. Making the driverstatic so that it does not open all those windows at once.

PC-325 Made the instance of WebDriver static so that only one browser window is used. Keeping it as protected instead of making private as I am lazy to add an accessor.